feat: display unification, audio/input vertical slices, CI, scaffolding
Display:
- Unify platform::DisplayDriver as single canonical trait (spec, update_buffer,
  refresh_fast, wait_ready); remove split with eink_emulator::DisplayDriver
- Unify EinkDisplay: add refresh_mode() getter and ghosting_level() to platform trait
- Fix EmulatorDisplay::update_buffer: unpack 2bpp framebuffer → DrawTarget pixels
- Fix Ssd1677Display::draw_iter: internal [u8; 96000] framebuffer, 2bpp MSB packing
- Fix LUT: remove wrong 133-byte table, use OTP LUT with comment
- Fix PanelType: Carta1300 → Carta1200
- Remove DapDisplay::update_buffer (now inherited from platform::DisplayDriver)

Audio:
- ES9038Q2M DAC driver: HAL-agnostic via embedded_hal_async::i2c::I2c<I>
- TPA6120A2 amp vertical slice: AmpDriver trait, Tpa6120a2<P: OutputPin>, MockAmp
- MockAmp: 6 tests (initial state, enable/disable, toggle, counts, default)

Input:
- Fluent InputBuilder API: rotary()/button() + .debounce_ms() + .emulated()/#[cfg]
- EmulatedAxis/EmulatedKey enums for keyboard mapping
- Wire spawn_input_task into main.rs (PE9/PE11 encoder, PD0-PD5 buttons)
- Keyboard input loop already wired in display_emulator.rs example

Platform:
- DMA busy-spin → embassy_futures::yield_now().await
- Hot-reload ui_version(): OnceLock<u64> with nanosecond timestamp (changes per load)
- Remove platform::mocks (unreachable module — use firmware-level mocks instead)

Workspace:
- Scaffold crates: playback, ui, library, bluetooth (no_std stubs + placeholder tests)
- Add memory.x linker script

CI:
- Rewrite .github/workflows/ci.yml: 5 jobs correctly feature-isolated
  (fmt, clippy, test, check-emulator, check-embedded for thumbv7em-none-eabihf)
- Add .pre-commit-config.yaml: fmt+clippy on commit, embedded check on push
- Add .github/dependabot.yml for GitHub Actions version tracking

+++ crates/eink/eink-emulator/examples/dap_display_demo.rs
@@ -103,23 +103,17 @@ fn draw_dap_ui(display: &mut Emulator) -> Result<(), Box<dyn std::error::Error>>
 
     // Album art placeholder (150x150, centered horizontally)
     let album_x = (800 - 150) / 2;
-    Rectangle::new(Point::new(album_x as i32, 80), Size::new(150, 150))
+    Rectangle::new(Point::new(album_x, 80), Size::new(150, 150))
         .into_styled(PrimitiveStyle::with_stroke(Gray4::BLACK, 2))
         .draw(display)?;
 
     // Diagonal cross for album placeholder
-    Line::new(
-        Point::new(album_x as i32, 80),
-        Point::new(album_x as i32 + 150, 230),
-    )
-    .into_styled(PrimitiveStyle::with_stroke(Gray4::new(12), 1))
-    .draw(display)?;
-    Line::new(
-        Point::new(album_x as i32 + 150, 80),
-        Point::new(album_x as i32, 230),
-    )
-    .into_styled(PrimitiveStyle::with_stroke(Gray4::new(12), 1))
-    .draw(display)?;
+    Line::new(Point::new(album_x, 80), Point::new(album_x + 150, 230))
+        .into_styled(PrimitiveStyle::with_stroke(Gray4::new(12), 1))
+        .draw(display)?;
+    Line::new(Point::new(album_x + 150, 80), Point::new(album_x, 230))
+        .into_styled(PrimitiveStyle::with_stroke(Gray4::new(12), 1))
+        .draw(display)?;
 
     // Album art label
     Text::new("Album Art", Point::new(350, 155), light_style).draw(display)?;
